+++ eyes-two/jni/hydrogen/jni/thresholder.cpp
@@ -22,26 +22,25 @@
 extern "C" {
 #endif  /* __cplusplus */
 
-jint Java_com_googlecode_eyesfree_textdetect_Thresholder_nativeSobelEdgeThreshold(JNIEnv *env,
-                                                                                  jclass clazz,
-                                                                                  jint nativePix,
-                                                                                  jint threshold) {
+jlong Java_com_googlecode_eyesfree_textdetect_Thresholder_nativeSobelEdgeThreshold(JNIEnv *env,
+                                                                                   jclass clazz,
+                                                                                   jlong nativePix,
+                                                                                   jlong threshold) {
   LOGV(__FUNCTION__);
 
   PIX *pixs = (PIX *) nativePix;
   PIX *pixd = pixThreshedSobelEdgeFilter(pixs, (l_int32) threshold);
 
-  return (jint) pixd;
+  return (jlong) pixd;
 }
 
-jint Java_com_googlecode_eyesfree_textdetect_Thresholder_nativeEdgeAdaptiveThreshold(
-                                                                                     JNIEnv *env,
-                                                                                     jclass clazz,
-                                                                                     jint nativePix,
-                                                                                     jint tileX,
-                                                                                     jint tileY,
-                                                                                     jint threshold,
-                                                                                     jint average) {
+jlong Java_com_googlecode_eyesfree_textdetect_Thresholder_nativeEdgeAdaptiveThreshold(JNIEnv *env,
+                                                                                      jclass clazz,
+                                                                                      jlong nativePix,
+                                                                                      jint tileX,
+                                                                                      jint tileY,
+                                                                                      jint threshold,
+                                                                                      jint average) {
   LOGV(__FUNCTION__);
 
   PIX *pixs = (PIX *) nativePix;
@@ -49,20 +48,19 @@ jint Java_com_googlecode_eyesfree_textdetect_Thresholder_nativeEdgeAdaptiveThres
 
   if (pixEdgeAdaptiveThreshold(pixs, &pixd, (l_int32) tileX, (l_int32) tileY, (l_int32) threshold,
                                (l_int32) average)) {
-    return (jint) 0;
+    return (jlong) 0;
   }
 
-  return (jint) pixd;
+  return (jlong) pixd;
 }
 
-jint Java_com_googlecode_eyesfree_textdetect_Thresholder_nativeFisherAdaptiveThreshold(
-                                                                                       JNIEnv *env,
-                                                                                       jclass clazz,
-                                                                                       jint nativePix,
-                                                                                       jint tileX,
-                                                                                       jint tileY,
-                                                                                       jfloat scoreFract,
-                                                                                       jfloat thresh) {
+jlong Java_com_googlecode_eyesfree_textdetect_Thresholder_nativeFisherAdaptiveThreshold(JNIEnv *env,
+                                                                                        jclass clazz,
+                                                                                        jlong nativePix,
+                                                                                        jint tileX,
+                                                                                        jint tileY,
+                                                                                        jfloat scoreFract,
+                                                                                        jfloat thresh) {
   LOGV(__FUNCTION__);
 
   PIX *pixs = (PIX *) nativePix;
@@ -70,10 +68,10 @@ jint Java_com_googlecode_eyesfree_textdetect_Thresholder_nativeFisherAdaptiveThr
 
   if (pixFisherAdaptiveThreshold(pixs, &pixd, (l_int32) tileX, (l_int32) tileY,
                                  (l_float32) scoreFract, (l_float32) thresh)) {
-    return (jint) 0;
+    return (jlong) 0;
   }
 
-  return (jint) pixd;
+  return (jlong) pixd;
 }
 
 #ifdef __cplusplus
