+++ tess-two/jni/com_googlecode_leptonica_android/box.cpp
@@ -20,48 +20,48 @@
 extern "C" {
 #endif  /* __cplusplus */
 
-jint Java_com_googlecode_leptonica_android_Box_nativeCreate(JNIEnv *env, jclass clazz, jint x,
-                                                            jint y, jint w, jint h) {
+jlong Java_com_googlecode_leptonica_android_Box_nativeCreate(JNIEnv *env, jclass clazz, jint x,
+                                                             jint y, jint w, jint h) {
   BOX *box = boxCreate((l_int32) x, (l_int32) y, (l_int32) w, (l_int32) h);
 
-  return (jint) box;
+  return (jlong) box;
 }
 
 void Java_com_googlecode_leptonica_android_Box_nativeDestroy(JNIEnv *env, jclass clazz,
-                                                             jint nativeBox) {
+                                                             jlong nativeBox) {
   BOX *box = (BOX *) nativeBox;
 
   boxDestroy(&box);
 }
 
-jint Java_com_googlecode_leptonica_android_Box_nativeGetX(JNIEnv *env, jclass clazz, jint nativeBox) {
+jint Java_com_googlecode_leptonica_android_Box_nativeGetX(JNIEnv *env, jclass clazz, jlong nativeBox) {
   BOX *box = (BOX *) nativeBox;
 
   return (jint) box->x;
 }
 
-jint Java_com_googlecode_leptonica_android_Box_nativeGetY(JNIEnv *env, jclass clazz, jint nativeBox) {
+jint Java_com_googlecode_leptonica_android_Box_nativeGetY(JNIEnv *env, jclass clazz, jlong nativeBox) {
   BOX *box = (BOX *) nativeBox;
 
   return (jint) box->y;
 }
 
 jint Java_com_googlecode_leptonica_android_Box_nativeGetWidth(JNIEnv *env, jclass clazz,
-                                                              jint nativeBox) {
+                                                              jlong nativeBox) {
   BOX *box = (BOX *) nativeBox;
 
   return (jint) box->w;
 }
 
 jint Java_com_googlecode_leptonica_android_Box_nativeGetHeight(JNIEnv *env, jclass clazz,
-                                                               jint nativeBox) {
+                                                               jlong nativeBox) {
   BOX *box = (BOX *) nativeBox;
 
   return (jint) box->h;
 }
 
 jboolean Java_com_googlecode_leptonica_android_Box_nativeGetGeometry(JNIEnv *env, jclass clazz,
-                                                                     jint nativeBox,
+                                                                     jlong nativeBox,
                                                                      jintArray dimensions) {
   BOX *box = (BOX *) nativeBox;
   jint *dimensionArray = env->GetIntArrayElements(dimensions, NULL);
