+++ tess-two/jni/com_googlecode_leptonica_android/pixa.cpp
@@ -20,37 +20,37 @@
 extern "C" {
 #endif  /* __cplusplus */
 
-jint Java_com_googlecode_leptonica_android_Pixa_nativeCreate(JNIEnv *env, jclass clazz, jint size) {
+jlong Java_com_googlecode_leptonica_android_Pixa_nativeCreate(JNIEnv *env, jclass clazz, jint size) {
   PIXA *pixa = pixaCreate((l_int32) size);
 
-  return (jint) pixa;
+  return (jlong) pixa;
 }
 
-jint Java_com_googlecode_leptonica_android_Pixa_nativeCopy(JNIEnv *env, jclass clazz,
-                                                           jint nativePixa) {
+jlong Java_com_googlecode_leptonica_android_Pixa_nativeCopy(JNIEnv *env, jclass clazz,
+                                                            jlong nativePixa) {
   PIXA *pixas = (PIXA *) nativePixa;
   PIXA *pixad = pixaCopy(pixas, L_CLONE);
 
-  return (jint) pixad;
+  return (jlong) pixad;
 }
 
-jint Java_com_googlecode_leptonica_android_Pixa_nativeSort(JNIEnv *env, jclass clazz,
-                                                           jint nativePixa, jint field, jint order) {
+jlong Java_com_googlecode_leptonica_android_Pixa_nativeSort(JNIEnv *env, jclass clazz,
+                                                            jlong nativePixa, jint field, jint order) {
   PIXA *pixas = (PIXA *) nativePixa;
   PIXA *pixad = pixaSort(pixas, field, order, NULL, L_CLONE);
 
-  return (jint) pixad;
+  return (jlong) pixad;
 }
 
 void Java_com_googlecode_leptonica_android_Pixa_nativeDestroy(JNIEnv *env, jclass clazz,
-                                                              jint nativePixa) {
+                                                              jlong nativePixa) {
   PIXA *pixa = (PIXA *) nativePixa;
 
   pixaDestroy(&pixa);
 }
 
 jboolean Java_com_googlecode_leptonica_android_Pixa_nativeJoin(JNIEnv *env, jclass clazz,
-                                                               jint nativePixa, jint otherPixa) {
+                                                               jlong nativePixa, jlong otherPixa) {
   PIXA *pixa = (PIXA *) nativePixa;
   PIXA *pixas = (PIXA *) otherPixa;
 
@@ -62,14 +62,14 @@ jboolean Java_com_googlecode_leptonica_android_Pixa_nativeJoin(JNIEnv *env, jcla
 }
 
 jint Java_com_googlecode_leptonica_android_Pixa_nativeGetCount(JNIEnv *env, jclass clazz,
-                                                               jint nativePixa) {
+                                                               jlong nativePixa) {
   PIXA *pixa = (PIXA *) nativePixa;
 
   return (jint) pixaGetCount(pixa);
 }
 
 void Java_com_googlecode_leptonica_android_Pixa_nativeAddPix(JNIEnv *env, jclass clazz,
-                                                             jint nativePixa, jint nativePix,
+                                                             jlong nativePixa, jlong nativePix,
                                                              jint mode) {
   PIXA *pixa = (PIXA *) nativePixa;
   PIX *pix = (PIX *) nativePix;
@@ -78,7 +78,7 @@ void Java_com_googlecode_leptonica_android_Pixa_nativeAddPix(JNIEnv *env, jclass
 }
 
 void Java_com_googlecode_leptonica_android_Pixa_nativeAddBox(JNIEnv *env, jclass clazz,
-                                                             jint nativePixa, jint nativeBox,
+                                                             jlong nativePixa, jlong nativeBox,
                                                              jint mode) {
   PIXA *pixa = (PIXA *) nativePixa;
   BOX *box = (BOX *) nativeBox;
@@ -87,8 +87,8 @@ void Java_com_googlecode_leptonica_android_Pixa_nativeAddBox(JNIEnv *env, jclass
 }
 
 void Java_com_googlecode_leptonica_android_Pixa_nativeAdd(JNIEnv *env, jclass clazz,
-                                                          jint nativePixa, jint nativePix,
-                                                          jint nativeBox, jint mode) {
+                                                          jlong nativePixa, jlong nativePix,
+                                                          jlong nativeBox, jint mode) {
   PIXA *pixa = (PIXA *) nativePixa;
   PIX *pix = (PIX *) nativePix;
   BOX *box = (BOX *) nativeBox;
@@ -98,8 +98,8 @@ void Java_com_googlecode_leptonica_android_Pixa_nativeAdd(JNIEnv *env, jclass cl
 }
 
 void Java_com_googlecode_leptonica_android_Pixa_nativeReplacePix(JNIEnv *env, jclass clazz,
-                                                                 jint nativePixa, jint index,
-                                                                 jint nativePix, jint nativeBox) {
+                                                                 jlong nativePixa, jint index,
+                                                                 jlong nativePix, jlong nativeBox) {
   PIXA *pixa = (PIXA *) nativePixa;
   PIX *pix = (PIX *) nativePix;
   BOX *box = (BOX *) nativeBox;
@@ -108,7 +108,7 @@ void Java_com_googlecode_leptonica_android_Pixa_nativeReplacePix(JNIEnv *env, jc
 }
 
 void Java_com_googlecode_leptonica_android_Pixa_nativeMergeAndReplacePix(JNIEnv *env, jclass clazz,
-                                                                         jint nativePixa,
+                                                                         jlong nativePixa,
                                                                          jint indexA, jint indexB) {
   PIXA *pixa = (PIXA *) nativePixa;
 
@@ -145,7 +145,7 @@ void Java_com_googlecode_leptonica_android_Pixa_nativeMergeAndReplacePix(JNIEnv 
 
 jboolean Java_com_googlecode_leptonica_android_Pixa_nativeWriteToFileRandomCmap(JNIEnv *env,
                                                                                 jclass clazz,
-                                                                                jint nativePixa,
+                                                                                jlong nativePixa,
                                                                                 jstring fileName,
                                                                                 jint width,
                                                                                 jint height) {
@@ -173,23 +173,23 @@ jboolean Java_com_googlecode_leptonica_android_Pixa_nativeWriteToFileRandomCmap(
 }
 
 jint Java_com_googlecode_leptonica_android_Pixa_nativeGetPix(JNIEnv *env, jclass clazz,
-                                                             jint nativePixa, jint index) {
+                                                             jlong nativePixa, jint index) {
   PIXA *pixa = (PIXA *) nativePixa;
   PIX *pix = pixaGetPix(pixa, (l_int32) index, L_CLONE);
 
-  return (jint) pix;
+  return (jlong) pix;
 }
 
-jint Java_com_googlecode_leptonica_android_Pixa_nativeGetBox(JNIEnv *env, jclass clazz,
-                                                             jint nativePixa, jint index) {
+jlong Java_com_googlecode_leptonica_android_Pixa_nativeGetBox(JNIEnv *env, jclass clazz,
+                                                              jlong nativePixa, jint index) {
   PIXA *pixa = (PIXA *) nativePixa;
   BOX *box = pixaGetBox(pixa, (l_int32) index, L_CLONE);
 
-  return (jint) box;
+  return (jlong) box;
 }
 
 jboolean Java_com_googlecode_leptonica_android_Pixa_nativeGetBoxGeometry(JNIEnv *env, jclass clazz,
-                                                                         jint nativePixa,
+                                                                         jlong nativePixa,
                                                                          jint index,
                                                                          jintArray dimensions) {
   PIXA *pixa = (PIXA *) nativePixa;
